Fix PandaX-4T S2-only limit (#34)
The plot of PandaX uses a  linear mass scale, not a logarithmic one, in figure 3 of https://arxiv.org/pdf/2507.11930. I checked a log scale when curve tracing, so the limit was only correct at the 2 and 6 GeV endpoints. At 3 GeV the limit is ~1e-43, whereas we had 4e-44 before.

Thanks to Genevieve Belanger for pointing out the error!